update
Now accepts both 1D and 2D spectra. Accepts list of input periods and list of lines. Option for standard output (only the data frame with the bandpasses that detect period) or full (everything).

Full output saves: plots of GLS and WF for each bandpass. dataframe with all bandpasses and periods. dataframe with bandpasses that have similar period to input. dataframe of equivalent_widths. plot of time series and GLS